pyproject.toml
[tool.poetry]
name = "gendo"
version = "0.1.0"
description = ""
authors = ["Иванов Владислав Сергеевич <ivanov_vs@rubin.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
jinja2 = "^3.1.2"
pyyaml = "^6.0"
click = "^8.1.3"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
